+++ editable.js
@@ -4128,6 +4128,7 @@ var content = (function() {
 
   var zeroWidthSpace = /\u200B/g;
   var zeroWidthNonBreakingSpace = /\uFEFF/g;
+  var whitespaceExceptSpace = /[^\S ]/g;
 
   return {
 
@@ -4181,6 +4182,10 @@ var content = (function() {
         element.removeChild(element.firstChild);
       }
       element.appendChild(fragment);
+    },
+
+    normalizeWhitespace: function(text) {
+      return text.replace(whitespaceExceptSpace, ' ');
     },
 
     /**
@@ -5585,13 +5590,29 @@ var highlightText = (function() {
 
   return {
     extractText: function(element) {
-      var textNode;
       var text = '';
-      var iterator = new NodeIterator(element);
-      while ( (textNode = iterator.getNextTextNode()) ) {
-        text = text + textNode.data;
-      }
+      this.getText(element, function(part) {
+        text += part;
+      });
       return text;
+    },
+
+    // Extract the text of an element.
+    // This has two notable behaviours:
+    // - It uses a NodeIterator which will skip elements
+    //   with data-editable="remove"
+    // - It returns a space for <br> elements
+    //   (The only block level element allowed inside of editables)
+    getText: function(element, callback) {
+      var iterator = new NodeIterator(element);
+      var next;
+      while ( (next = iterator.getNext()) ) {
+        if (next.nodeType === nodeType.textNode && next.data !== '') {
+          callback(next.data);
+        } else if (next.nodeType === nodeType.elementNode && next.nodeName === 'BR') {
+          callback(' ');
+        }
+      }
     },
 
     highlight: function(element, regex, stencilElement) {
@@ -6879,6 +6900,8 @@ var Spellcheck = (function() {
   Spellcheck.prototype.checkSpelling = function(editableHost) {
     var that = this;
     var text = highlightText.extractText(editableHost);
+    text = content.normalizeWhitespace(text);
+
     this.config.spellcheckService(text, function(misspelledWords) {
       var selection = that.editable.getSelection(editableHost);
       if (selection) {