+++ Changelog.md
@@ -1,3 +1,9 @@
+# v0.4.2
+
+#### Bugfixes
+
+- Fix spellcheck whitespace handling [#87](https://github.com/upfrontIO/editable.js/pull/87)
+
 # v0.4.1
 
 #### Features